Made things neat
